feat: move `FileAssetId` to correct folder and update using statements accoring to namespace change
