+++ Backend/src/BuildingBlocks/BuildingBlocks.Api/Converters/FileIdConverter.cs
@@ -1,4 +1,4 @@
-﻿using BuildingBlocks.Domain.ValueObjects.Ids;
+﻿using BuildingBlocks.Domain.Files.File.ValueObjects;
 
 namespace BuildingBlocks.Api.Converters;
