+++ Backend/src/BuildingBlocks/BuildingBlocks.Domain/BuildingBlocks.Domain.csproj
@@ -11,8 +11,4 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.0" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Files\File\ValueObjects\" />
-  </ItemGroup>
-
 </Project>
