+++ Backend/src/Modules/Files/Files.Api/Endpoints/Files/CreateFileAsset.cs
@@ -1,5 +1,5 @@
 ﻿using BuildingBlocks.Domain.Files.File.Dtos;
-using BuildingBlocks.Domain.ValueObjects.Ids;
+using BuildingBlocks.Domain.Files.File.ValueObjects;
 using Files.Application.Entities.Files.Commands.CreateFileAsset;
 
 // ReSharper disable ClassNeverInstantiated.Global
