+++ Backend/src/Modules/Files/Files.Application/Entities/Files/Commands/CreateFileAsset/CreateFileAssetHandler.cs
@@ -1,4 +1,6 @@
-﻿namespace Files.Application.Entities.Files.Commands.CreateFileAsset;
+﻿using BuildingBlocks.Domain.Files.File.ValueObjects;
+
+namespace Files.Application.Entities.Files.Commands.CreateFileAsset;
 
 internal class CreateFileAssetHandler(IFilesDbContext dbContext) :
     ICommandHandler<CreateFileAssetCommand, CreateFileAssetResult>
