+++ Backend/src/Modules/Files/Files.Application/Entities/Files/Queries/GetFileAssetById/GetFileAssetByIdQuery.cs
@@ -1,6 +1,7 @@
 ﻿// ReSharper disable ClassNeverInstantiated.Global
 
 using BuildingBlocks.Domain.Files.File.Dtos;
+using BuildingBlocks.Domain.Files.File.ValueObjects;
 
 namespace Files.Application.Entities.Files.Queries.GetFileAssetById;
